+++ pyunfold/mix.py
@@ -9,9 +9,9 @@ from .utils import safe_inverse
 class Mixer(object):
     """DAgostini Bayesian Mixer Class
     """
-    def __init__(self, error_type='ACM', MCTables=None, data=None,
-                 data_err=None, efficiencies=None, efficiencies_err=None,
-                 response=None, response_err=None):
+    def __init__(self, data=None, data_err=None, efficiencies=None,
+                 efficiencies_err=None, response=None, response_err=None,
+                 cov_type='multinomial'):
         # Input validation
         if len(data) != response.shape[0]:
             err_msg = ('Inconsistent number of effect bins. Observed data '
@@ -23,7 +23,6 @@ class Mixer(object):
         self.pec = response
         self.cEff = efficiencies
         self.NEobs = data
-        self.error_type = error_type
 
         # Number of Cause and Effect Bins
         dims = self.pec.shape
@@ -33,13 +32,13 @@ class Mixer(object):
         # Mixing Matrix
         self.Mij = np.zeros(dims)
 
-        self.cov = CovarianceMatrix(error_type=error_type,
-                                    data=data,
+        self.cov = CovarianceMatrix(data=data,
                                     data_err=data_err,
                                     efficiencies=efficiencies,
                                     efficiencies_err=efficiencies_err,
                                     response=response,
-                                    response_err=response_err)
+                                    response_err=response_err,
+                                    cov_type=cov_type)
 
     def get_cov(self):
         """Covariance Matrix
@@ -103,18 +102,18 @@ class CovarianceMatrix(object):
     (http://hepunx.rl.ac.uk/~adye/software/unfold/RooUnfold.html#dev),
     or based on derivations presented in Unfolding reference section 'Unfolding Uncertainties'
     """
-    def __init__(self, error_type='ACM', MCTables=None, data=None, data_err=None,
-                 efficiencies=None, efficiencies_err=None, response=None,
-                 response_err=None):
+    def __init__(self, data=None, data_err=None, efficiencies=None,
+                 efficiencies_err=None, response=None, response_err=None,
+                 cov_type='multinomial'):
 
-        self.error_type = error_type
         # Normalized P(E|C)
         self.pec = response
         self.pec_err = response_err
         # Errors on P(E|C), for Vc1
-        self.pec_cov_type = "Multinomial"
-        # self.pec_cov_type = "Poisson"
-        self.PecCov = self.SetPecCovType(self.pec_cov_type)
+        if cov_type.lower() not in ['multinomial', 'poisson']:
+            raise ValueError('Invalid pec_cov_type entered: {}. Must be '
+                             'either "multinomial" or "poisson".'.format(cov_type))
+        self.pec_cov_type = cov_type.lower()
         self.cEff = efficiencies
         self.cEff_inv = safe_inverse(self.cEff)
         # Effective number of sim events
@@ -132,8 +131,6 @@ class CovarianceMatrix(object):
         self.ebins = dims[0]
         # Mixing matrix
         self.Mij = np.zeros(dims)
-        # Flag to propagate derivative
-        self.ErrorPropFlag = self.SetErrorPropFlag(self.error_type)
         # Adye propagating derivative
         self.dcdn = np.zeros(dims)
         self.dcdP = np.zeros((self.cbins, self.cbins * self.ebins))
@@ -166,7 +163,7 @@ class CovarianceMatrix(object):
             dcdP[ti, ec_j+ti] += (n_c_prev[ti] * NE_F_R[ej] - n_c[ti]) * self.cEff_inv[ti]
 
         # Adye propagation corrections
-        if self.ErrorPropFlag and self.counter > 0:
+        if self.counter > 0:
             # Get previous derivatives
             dcdn_prev = self.dcdn.copy()
             dcdP_prev = self.dcdP.copy()
@@ -226,23 +223,21 @@ class CovarianceMatrix(object):
         CovPP = np.zeros((cbins * ebins, cbins * ebins))
 
         # Poisson covariance matrix
-        if self.PecCov == 1:
+        if self.pec_cov_type == 'poisson':
             for ej, ti in product(range(0, ebins), range(0, cbins)):
                 ejc = ej * cbins
                 CovPP[ejc+ti, ejc+ti] = self.pec_err[ej, ti]**2
         # Multinomial covariance matrix
-        elif self.PecCov == 2:
-            NC_inv = safe_inverse(self.NCmc)
+        elif self.pec_cov_type == 'multinomial':
+            nc_inv = safe_inverse(self.NCmc)
             for ej, ti in product(range(0, ebins), range(0, cbins)):
                 ejc = ej * cbins
                 # Don't go looping if zeros are present :)
-                if (self.pec[ej, ti] > 0 and NC_inv[ti] > 0):
-                    CovPP[ejc+ti, ejc+ti] = NC_inv[ti] * self.pec[ej, ti] * (1 - self.pec[ej, ti])
-                    cov1 = -NC_inv[ti] * self.pec[ej, ti]
+                if (self.pec[ej, ti] > 0 and nc_inv[ti] > 0):
+                    CovPP[ejc+ti, ejc+ti] = nc_inv[ti] * self.pec[ej, ti] * (1 - self.pec[ej, ti])
+                    cov1 = -nc_inv[ti] * self.pec[ej, ti]
                     for ek in range(ej+1, ebins):
                         ekc = ek * cbins
-                        if ejc+ti == ekc+ti or ek == ej:
-                            continue
                         cov = cov1 * self.pec[ek, ti]
                         CovPP[ejc+ti, ekc+ti] = cov
                         CovPP[ekc+ti, ejc+ti] = cov
@@ -269,33 +264,3 @@ class CovarianceMatrix(object):
         # Full Covariance Matrix
         Vc = Vc0 + Vc1
         return Vc
-
-    def SetErrorPropFlag(self, func):
-        """Key to choose whether to propagate errors at each iteration
-        """
-        # ACM propagate errors as Adye
-        # DCM propagate errors as DAgostini
-        ErrorPropOptionsKey = {
-            "ACM": True,
-            "DCM": False,
-            }
-        if func in ErrorPropOptionsKey:
-            return ErrorPropOptionsKey[func]
-        else:
-            err_msg = ('Invalid error propagation type entered, {}, must '
-                       'be in {}'.format(func, ErrorPropOptionsKey.keys()))
-            raise ValueError(err_msg)
-
-    def SetPecCovType(self, func):
-        """Key for Selecting PEC Covariance Matrix Type
-        """
-        PecCovOptionsKey = {
-            "Poisson": 1,
-            "Multinomial": 2
-            }
-        if func in PecCovOptionsKey:
-            return PecCovOptionsKey[func]
-        else:
-            err_msg = ('Invalid covariance calculation type entered, {}, must '
-                       'be in {}'.format(func, PecCovOptionsKey.keys()))
-            raise ValueError(err_msg)