+++ pyunfold/tests/test_unfold.py
@@ -9,6 +9,7 @@ import pytest
 from .testing_utils import diagonal_response, triangular_response
 
 from pyunfold.unfold import iterative_unfold
+from pyunfold.priors import jeffreys_prior
 
 PY_VERSION = sys.version_info.major
 
@@ -35,7 +36,6 @@ def test_iterative_unfold(response_type):
     unfolded_result = iterative_unfold(counts, counts_err,
                                        response, response_err,
                                        efficiencies, efficiencies_err,
-                                       priors='Jeffreys',
                                        ts='ks',
                                        ts_stopping=0.01,
                                        return_iterations=False)
@@ -53,20 +53,23 @@ def test_iterative_unfold_max_iter():
     # Load test counts distribution and diagonal response matrix
     np.random.seed(2)
 
-    samples = np.random.normal(loc=0, scale=1, size=int(1e5))
+    samples = np.random.normal(loc=10, scale=1, size=int(1e5))
 
-    bins = np.linspace(-1, 1, 10)
+    bins = np.linspace(8, 12, 10)
     counts, _ = np.histogram(samples, bins=bins)
     counts_err = np.sqrt(counts)
     response, response_err = triangular_response(len(counts))
     efficiencies = np.ones_like(counts, dtype=float)
     efficiencies_err = np.full_like(efficiencies, 0.001)
 
+    causes = (bins[1:] + bins[:-1]) / 2
+    prior = jeffreys_prior(causes=causes)
+
     max_iter = 5
     unfolded_result = iterative_unfold(counts, counts_err,
                                        response, response_err,
                                        efficiencies, efficiencies_err,
-                                       priors='Jeffreys',
+                                       prior=prior,
                                        ts='ks',
                                        ts_stopping=0.00001,
                                        max_iter=max_iter,
@@ -92,10 +95,17 @@ def test_example():
                     [0.01, 0.01]]
     efficiencies = [0.4, 0.67]
     efficiencies_err = [0.01, 0.01]
+    causes = np.arange(len(data)) + 0.5
+    prior = jeffreys_prior(causes=causes)
+
     # Perform iterative unfolding
-    unfolded = iterative_unfold(data, data_err,
-                                response, response_err,
-                                efficiencies, efficiencies_err,
+    unfolded = iterative_unfold(data=data,
+                                data_err=data_err,
+                                response=response,
+                                response_err=response_err,
+                                efficiencies=efficiencies,
+                                efficiencies_err=efficiencies_err,
+                                prior=prior,
                                 return_iterations=True)
     unfolded = unfolded[columns]
 
@@ -119,12 +129,12 @@ def test_example_2():
                     [0.01, 0.01]]
     efficiencies = [0.4, 0.67]
     efficiencies_err = [0.01, 0.01]
-    priors = [0.34, 1-0.34]
+    prior = [0.34, 1-0.34]
     # Perform iterative unfolding
     unfolded = iterative_unfold(data, data_err,
                                 response, response_err,
                                 efficiencies, efficiencies_err,
-                                priors=priors,
+                                prior=prior,
                                 return_iterations=True)
     unfolded = unfolded[columns]
 
@@ -148,12 +158,12 @@ def test_example_non_square_response():
                     [0.01, 0.01, 0.01]]
     efficiencies = [0.4, 0.67, 0.8]
     efficiencies_err = [0.01, 0.01, 0.01]
-    priors = [0.34, 0.21, 1 - (0.34 + 0.21)]
+    prior = [0.34, 0.21, 1 - (0.34 + 0.21)]
     # Perform iterative unfolding
     unfolded = iterative_unfold(data, data_err,
                                 response, response_err,
                                 efficiencies, efficiencies_err,
-                                priors=priors,
+                                prior=prior,
                                 return_iterations=True)
     unfolded = unfolded[columns]
 
@@ -164,27 +174,54 @@ inputs = ['data', 'data_err', 'response', 'response_err', 'efficiencies', 'effic
 
 
 @pytest.mark.parametrize('none_input', inputs)
-def test_iterative_unfold_none_input_raises(none_input):
-    # Load test counts distribution and diagonal response matrix
-    np.random.seed(2)
-    samples = np.random.normal(loc=0, scale=1, size=int(1e5))
-
-    bins = np.linspace(-1, 1, 10)
-    data, _ = np.histogram(samples, bins=bins)
-    data_err = np.sqrt(data)
-    response, response_err = diagonal_response(len(data))
-    efficiencies = np.ones_like(data, dtype=float)
-    efficiencies_err = np.full_like(efficiencies, 0.001)
-
-    inputs = {'data': data,
-              'data_err': data_err,
-              'response': response,
-              'response_err': response_err,
-              'efficiencies': efficiencies,
-              'efficiencies_err': efficiencies_err
+def test_iterative_unfold_none_input_raises(none_input, example_dataset):
+    inputs = {'data': example_dataset.data,
+              'data_err': example_dataset.data_err,
+              'response': example_dataset.response,
+              'response_err': example_dataset.response_err,
+              'efficiencies': example_dataset.efficiencies,
+              'efficiencies_err': example_dataset.efficiencies_err
               }
     inputs[none_input] = None
     with pytest.raises(ValueError) as excinfo:
         iterative_unfold(**inputs)
     expected_msg = 'The input for "{}" must not be None.'.format(none_input)
+    assert expected_msg == str(excinfo.value)
+
+
+@pytest.mark.parametrize('cov_type_1,cov_type_2', [('multinomial', 'Multinomial'),
+                                                   ('poisson', 'Poisson')])
+def test_iterative_unfold_cov_type_case_insensitive(cov_type_1, cov_type_2, example_dataset):
+    # Test that cov_type is case insensitive
+    inputs = {'data': example_dataset.data,
+              'data_err': example_dataset.data_err,
+              'response': example_dataset.response,
+              'response_err': example_dataset.response_err,
+              'efficiencies': example_dataset.efficiencies,
+              'efficiencies_err': example_dataset.efficiencies_err
+              }
+    result_1 = iterative_unfold(cov_type=cov_type_1, **inputs)
+    result_2 = iterative_unfold(cov_type=cov_type_2, **inputs)
+
+    assert result_1.keys() == result_2.keys()
+    for key in result_1.keys():
+        if isinstance(result_1[key], np.ndarray):
+            np.testing.assert_array_equal(result_1[key], result_2[key])
+        else:
+            assert result_1[key] == result_2[key]
+
+
+def test_iterative_unfold_cov_type_raises(example_dataset):
+    inputs = {'data': example_dataset.data,
+              'data_err': example_dataset.data_err,
+              'response': example_dataset.response,
+              'response_err': example_dataset.response_err,
+              'efficiencies': example_dataset.efficiencies,
+              'efficiencies_err': example_dataset.efficiencies_err
+              }
+    cov_type = 'Not a valid cov_type'
+    with pytest.raises(ValueError) as excinfo:
+        iterative_unfold(cov_type=cov_type, **inputs)
+    expected_msg = ('Invalid pec_cov_type entered: {}. Must be '
+                    'either "multinomial" or "poisson".'.format(cov_type))
     assert expected_msg == str(excinfo.value)