+++ pyunfold/unfold.py
@@ -12,9 +12,9 @@ from .callbacks import validate_callbacks, extract_regularizer
 
 def iterative_unfold(data=None, data_err=None, response=None,
                      response_err=None, efficiencies=None,
-                     efficiencies_err=None, priors='Jeffreys', ts='ks',
-                     ts_stopping=0.01, max_iter=100, return_iterations=False,
-                     callbacks=None):
+                     efficiencies_err=None, prior=None, ts='ks',
+                     ts_stopping=0.01, max_iter=100, cov_type='multinomial',
+                     return_iterations=False, callbacks=None):
     """Performs iterative Bayesian unfolding
 
     Parameters
@@ -34,10 +34,10 @@ def iterative_unfold(data=None, data_err=None, response=None,
     efficiencies_err : array_like
         Uncertainties of detection efficiencies. Must be the same shape as
         ``efficiencies``.
-    priors : str or array_like, optional
-        Prior distribution to use in unfolding. If 'Jeffreys', then the
-        Jeffreys (flat) prior is used. Otherwise, must be array_like with
-        same shape as ``efficiencies`` (default is 'Jeffreys').
+    prior : array_like, optional
+        Prior distribution to use in unfolding. If None, then a uniform
+        (or flat) prior will be used. If array_like, then must have the same
+        shape as ``efficiencies`` (default is None).
     ts : {'ks', 'chi2', 'pf', 'rmd'}
         Name of test statistic to use for stopping condition (default is 'ks').
     ts_stopping : float, optional
@@ -47,6 +47,9 @@ def iterative_unfold(data=None, data_err=None, response=None,
         procedure is stopped (default is 0.01).
     max_iter : int, optional
         Maximum number of iterations to allow (default is 100).
+    cov_type : {'multinomial', 'poisson'}
+        Whether to use the Multinomial or Poisson form for the covariance
+        matrix (default is 'multinomial').
     return_iterations : bool, optional
         Whether to return unfolded distributions for each iteration
         (default is False).
@@ -105,18 +108,19 @@ def iterative_unfold(data=None, data_err=None, response=None,
     num_causes = len(efficiencies)
 
     # Setup prior
-    n_c = setup_prior(priors=priors,
-                      num_causes=num_causes,
-                      num_observations=np.sum(data))
+    prior = setup_prior(prior=prior, num_causes=num_causes)
+
+    # Define first prior counts distribution
+    n_c = np.sum(data) * prior
 
     # Setup Mixer
-    mixer = Mixer(error_type='ACM',
-                  data=data,
+    mixer = Mixer(data=data,
                   data_err=data_err,
                   efficiencies=efficiencies,
                   efficiencies_err=efficiencies_err,
                   response=response,
-                  response_err=response_err)
+                  response_err=response_err,
+                  cov_type=cov_type)
 
     # Setup test statistic
     ts_obj = get_ts(ts)